rename: 함수와 컴포넌트 분리
