+++ client/src/components/compete/CompDetail.js
@@ -2,8 +2,8 @@ import { useLocation } from "react-router-dom";
 import styled from "styled-components";
 import BackHeader from "components/common/BackHeader";
 import { Count, Date } from "./comp/Card";
-import kFormatter from "./comp/kFormatter";
-import formatDate from "./comp/DateChanger";
+import kFormatter from "./function/kFormatter";
+import formatDate from "./function/DateChanger";
 import PercentBar from "./comp/PercentBar";
 import React, { useState } from "react";
 import { Timer, TimerBox } from "./comp/Timer";
