+++ client/src/components/compete/function/kFormatter.js
